chore: move removeStablecoinsFromSourceTokens util into hook
