+++ app/components/UI/Bridge/Views/BatchSellTokenSelect/BatchSellTokenSelect.utils.ts
@@ -1,9 +1,8 @@
 import {
-  formatAddressToAssetId,
   formatChainIdToCaip,
   formatChainIdToHex,
 } from '@metamask/bridge-controller';
-import { CaipAssetType, CaipChainId } from '@metamask/utils';
+import { CaipChainId } from '@metamask/utils';
 import { NETWORK_TO_SHORT_NETWORK_NAME_MAP } from '../../../../../constants/bridge';
 import { BridgeToken } from '../../types';
 
@@ -41,44 +40,6 @@ export function getBatchSellDestinationToken(
   return stablecoins.find(
     (stablecoin) => formatChainIdToCaip(stablecoin.chainId) === caipChainId,
   );
-}
-
-export function removeStablecoinsFromSourceTokens({
-  tokens,
-  stablecoinsByChain,
-}: {
-  tokens: BridgeToken[];
-  stablecoinsByChain: Partial<Record<CaipChainId, BridgeToken[]>>;
-}): BridgeToken[] {
-  const stablecoinAssetIdsByChain = new Map(
-    Object.entries(stablecoinsByChain).map(([chainId, stablecoins]) => [
-      chainId as CaipChainId,
-      new Set(
-        (stablecoins ?? [])
-          .map((stablecoin) =>
-            formatAddressToAssetId(stablecoin.address, stablecoin.chainId),
-          )
-          .filter((assetId): assetId is CaipAssetType => Boolean(assetId)),
-      ),
-    ]),
-  );
-
-  return tokens.filter((token) => {
-    const caipChainId = formatChainIdToCaip(token.chainId);
-    const stablecoinAssetIds = stablecoinAssetIdsByChain.get(caipChainId);
-
-    if (!stablecoinAssetIds) {
-      return true;
-    }
-
-    const assetId = formatAddressToAssetId(token.address, token.chainId);
-
-    if (!assetId) {
-      return true;
-    }
-
-    return !stablecoinAssetIds.has(assetId);
-  });
 }
 
 export function sortBatchSellTokens(
